+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.37
+
+* `Dialog` now do not add auto margin by default
+* Support of `Accordion` element
+
 ## 0.0.36
 
 * Reorder arguments in `DefaultComment` fun